requirements.txt
addict==2.4.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bleach==4.1.0
certifi==2022.6.15
cffi==1.15.1
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.17.2
Jinja2==3.0.3
json5==0.9.9
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
open3d-python==0.7.0.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pyzmq==23.2.0
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.6.0
tomlkit==0.11.4
torch==1.10.2
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
zipp==3.6.0
